+++ kinetics/__init__.py
@@ -1,19 +1,7 @@
-from kinetics.Model import Model, uM_to_mgml
-
+from kinetics.Model import Model, Metrics, uM_to_mgml
 from kinetics.Reaction_Classes import *
-
-from kinetics.Uncertainty_Sensitivity_Analysis import UA, SA, get_bounds_from_pc_error, get_bounds_from_std_error
-
-from kinetics.graph_functions import setup_graph, add_ua_model_to_graph, add_experimental_data_to_graph, remove_st_less_than, plot_sa_total_sensitivity, set_graph_settings, plot_all_runs
-
-from kinetics.Initial_Rates import calc_initial_rates_ua, fit_mm, concentrations_around_km,  calc_initial_rates_single, plot_fit_ua, plot_scatter_all_runs
-
-from kinetics.ga import GA_Base_Class
-
-from kinetics.metrics import Metrics
-
 from kinetics import Uncertainty, Sensitivity
+from kinetics.Initial_Rates import *
+from kinetics.Genetic_Algorithm import GA_Base_Class
 
-from kinetics import Enzymes
-
-__version__ = '1.2.3'
+__version__ = '1.2.4'
